requirements.txt
streamlit
torch
transformers 
matplotlib
pandas 
selenium 
google-generativeai